+++ src/components/accordion-item/resources.ts
@@ -1,0 +1,17 @@
+export const SLOTS = {
+  actionsStart: "actions-start",
+  actionsEnd: "actions-end"
+};
+
+export const CSS = {
+  icon: "icon",
+  header: "header",
+  headerContent: "header-content",
+  actionsStart: "actions-start",
+  actionsEnd: "actions-end",
+  headerText: "header-text",
+  heading: "heading",
+  description: "description",
+  expandIcon: "expand-icon",
+  content: "content"
+};
